--- test/d2l-organization-image/d2l-organization-image.html
@@ -1,31 +1,0 @@
-<html>
-	<head>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes">
-		<script src="../../../@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
-		<script src="../../../wct-browser-legacy/browser.js"></script>
-
-		<script type="module" src="../../components/d2l-organization-image/d2l-organization-image.js"></script>
-	</head>
-	<body>
-		<test-fixture id="no-params">
-			<template>
-				<d2l-organization-image></d2l-organization-image>
-			</template>
-		</test-fixture>
-
-		<test-fixture id="with-href">
-			<template>
-				<d2l-organization-image href="/organizations/1" token="whatever"></d2l-organization-image>
-			</template>
-		</test-fixture>
-
-		<test-fixture id="with-learning-paths">
-				<template>
-					<d2l-organization-image href="/learningPaths/1" token="whatever2"></d2l-organization-image>
-				</template>
-			</test-fixture>
-
-		<script type="module" src="./d2l-organization-image.js"></script>
-	</body>
-</html>
