+++ test/index.html
@@ -8,7 +8,6 @@
 	<body>
 		<script>
 			WCT.loadSuites([
-				'./d2l-organization-image/d2l-organization-image.html',
 				'./d2l-organization-info/d2l-organization-info.html',
 				'./d2l-organization-name/d2l-organization-name.html',
 				'./d2l-organization-updates/d2l-organization-updates.html',
